refactor: Update QuantumState constructor to accept data and dims parameters
The QuantumState constructor in the QuantumState class has been updated to accept the data and dims parameters. This change allows for more flexibility in initializing a QuantumState object with different types of data, such as numpy arrays, lists, Statevectors, Operators, QuantumCircuits, or Instructions. The dims parameter can be used to specify the dimensions of the quantum state. This enhancement improves the usability and versatility of the QuantumState class.