refactor: Update draw method in QuantumState class
The draw method in the QuantumState class has been updated to provide additional visualization options for the quantum state vector. This includes the ability to output the state vector as a LaTeX formatted matrix or vector. The update also allows for specifying the target basis for visualization and controlling the display of qubit indices. These enhancements improve the flexibility and usability of the draw method.